backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.12.0
langchain-openai==1.1.0
langchain-core==0.1.23
pydantic==2.12.5